Moved parley-network module back to the main parley module
