+++ parley/build.gradle
@@ -59,8 +59,9 @@ dependencies {
     // Android Material
     implementation 'com.google.android.material:material:1.11.0'
 
-    // Gson
-    implementation "com.google.code.gson:gson:2.11.0"
+    // Retrofit
+    implementation "com.squareup.retrofit2:retrofit:2.9.0"
+    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
 
     // Glide
     implementation "com.github.bumptech.glide:glide:4.15.1"
