+++ parley/src/main/java/nu/parley/android/NetworkConfig.kt
@@ -1,6 +1,5 @@
 package nu.parley.android
 
-import nu.parley.android.data.net.ParleyJsonParser
 import nu.parley.android.data.net.ParleyRepositories
 
 interface NetworkConfig {
