+++ parley/src/main/java/nu/parley/android/ParleyNetworkConfig.kt
@@ -1,0 +1,10 @@
+package nu.parley.android
+
+import nu.parley.android.data.repository.DefaultParleyRepositories
+import okhttp3.Interceptor
+
+class ParleyNetworkConfig(
+    val interceptor: Interceptor? = null
+) : NetworkConfig {
+    override val repositories = DefaultParleyRepositories
+}
